feat(islands):
* refactor record for Island coordinates x and y as double
* refactor IslandComponent.fxml. Now it's a pane with 2 images and the possibility to show a 'flag' over the island
* update IslandComponent.java

close #296, #300

+++ src/main/java/de/uniks/stp24/model/Island.java
@@ -3,8 +3,8 @@ package de.uniks.stp24.model;
 public record Island(
   String owner,
   int flagIndex,
-  int posX,
-  int posY,
+  double posX,
+  double posY,
   IslandType type,
   int crewCapacity,
   int resourceCapacity,
